Only allow the customer set the balance
